Refactored the malaria cnn scripts under a new malaria.cnn module and created a malaria.yolo model for the new model
