+++ malaria/cnn/model.py
